% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_rmd.R
\name{output_rmd}
\alias{output_rmd}
\title{A output_Rmd Function}
\usage{
output_rmd(input_values, file_name, filename_extension)
}
\arguments{
\item{input_values}{Data of type list.}

\item{file_name}{file name}

\item{filename_extension}{filename extension}
}
\description{
This function can export the contents of the list to the Rmarkdown document.
}
\examples{
output_rmd()
}
